shared_library.strings module docs
